In CreateAccountActivity, edited to create a proper user profile with their first name and last name also including other side information such as their follower count. This is not final, it is a place holder for their information format.
Login Activity also improved to display proper error messages such as the database not findind the proper combination to the
user info.

TrendingActivity and its associated xml file changed to implement a button to start the UserPageActivity.

User class built in order to build a proper template for the user's information.